test_agent.py
import os
import sys

sys.path.append(os.getcwd())

import google.generativeai as genai
from src.config import settings

genai.configure(api_key=settings.gemini_api_key)

print("Available models:")
for m in genai.list_models():
    if 'generateContent' in m.supported_generation_methods:
        print(m.name)
